move Display components into own dir
