+++ src/components/displays/OscilloscopeChart.tsx
