+++ src/components/displays/Rocket.tsx
