+++ src/components/displays/SolidChart.tsx
