+++ src/components/displays/TemplateDisplayComponent.tsx
